Move eval to internal/eval
