+++ internal/eval/eval.go
@@ -1,4 +1,4 @@
-package runn
+package eval
 
 import (
 	"fmt"
@@ -15,8 +15,8 @@ import (
 	"github.com/goccy/go-yaml"
 	"github.com/k1LoW/expand"
 	"github.com/k1LoW/runn/builtin"
-	"github.com/k1LoW/runn/internal/exprtrace"
 	"github.com/k1LoW/runn/internal/deprecation"
+	"github.com/k1LoW/runn/internal/exprtrace"
 	"github.com/xlab/treeprint"
 )
