+++ internal/eval/eval_test.go
@@ -1,4 +1,4 @@
-package runn
+package eval
 
 import (
 	"fmt"
